+++ src/middleware/prepareOutput.js
@@ -1,6 +1,5 @@
 'use strict'
 
-const _ = require('lodash')
 const asyncEachSeries = require('async/eachSeries')
 
 module.exports = function (options, excludedMap) {
@@ -48,13 +47,19 @@ module.exports = function (options, excludedMap) {
       }
 
       if (options.totalCountHeader && req.erm.totalCount) {
-        res.header(_.isString(options.totalCountHeader) ? options.totalCountHeader : 'X-Total-Count', req.erm.totalCount)
+        res.header(typeof options.totalCountHeader === 'string' ? options.totalCountHeader : 'X-Total-Count', req.erm.totalCount)
       }
 
-      options.outputFn(req, res)
+      const promise = options.outputFn(req, res)
 
       if (options.postProcess) {
-        options.postProcess(req, res, next)
+        if (promise && typeof promise.then === 'function') {
+          promise.then(() => {
+            options.postProcess(req, res, next)
+          }).catch(errorHandler(req, res, next))
+        } else {
+          options.postProcess(req, res, next)
+        }
       }
     })
   }