+++ test/integration/middleware.js
@@ -40,9 +40,7 @@ module.exports = function (createFn, setup, dismantle) {
         }).then((createdCustomer) => {
           customer = createdCustomer
           server = app.listen(testPort, done)
-        }, (err) => {
-          done(err)
-        })
+        }).catch(done)
       })
     })
 
@@ -125,9 +123,7 @@ module.exports = function (createFn, setup, dismantle) {
         }).then((createdCustomer) => {
           customer = createdCustomer
           server = app.listen(testPort, done)
-        }, (err) => {
-          done(err)
-        })
+        }).catch(done)
       })
     })
 
@@ -366,9 +362,7 @@ module.exports = function (createFn, setup, dismantle) {
         }).then((createdCustomer) => {
           customer = createdCustomer
           server = app.listen(testPort, done)
-        }, (err) => {
-          done(err)
-        })
+        }).catch(done)
       })
     })
 
@@ -470,9 +464,7 @@ module.exports = function (createFn, setup, dismantle) {
         }).then((createdCustomer) => {
           customer = createdCustomer
           server = app.listen(testPort, done)
-        }, (err) => {
-          done(err)
-        })
+        }).catch(done)
       })
     })
 
@@ -558,9 +550,7 @@ module.exports = function (createFn, setup, dismantle) {
         }).then((createdCustomer) => {
           customer = createdCustomer
           server = app.listen(testPort, done)
-        }, (err) => {
-          done(err)
-        })
+        }).catch(done)
       })
     })
 
@@ -628,9 +618,7 @@ module.exports = function (createFn, setup, dismantle) {
         }).then((createdCustomer) => {
           customer = createdCustomer
           server = app.listen(testPort, done)
-        }, (err) => {
-          done(err)
-        })
+        }).catch(done)
       })
     })
 
@@ -793,9 +781,7 @@ module.exports = function (createFn, setup, dismantle) {
         }).then((createdCustomer) => {
           customer = createdCustomer
           server = app.listen(testPort, done)
-        }, (err) => {
-          done(err)
-        })
+        }).catch(done)
       })
     })
 
@@ -921,9 +907,7 @@ module.exports = function (createFn, setup, dismantle) {
         }).then((createdCustomer) => {
           customer = createdCustomer
           server = app.listen(testPort, done)
-        }, (err) => {
-          done(err)
-        })
+        }).catch(done)
       })
     })
 
@@ -1037,9 +1021,7 @@ module.exports = function (createFn, setup, dismantle) {
         }).then((createdCustomer) => {
           customer = createdCustomer
           server = app.listen(testPort, done)
-        }, (err) => {
-          done(err)
-        })
+        }).catch(done)
       })
     })
 
@@ -1161,6 +1143,60 @@ module.exports = function (createFn, setup, dismantle) {
     let app = createFn()
     let server
     let options = {
+      postProcess: sinon.spy((req, res, next) => {
+        next()
+      }),
+      restify: app.isRestify
+    }
+
+    beforeEach((done) => {
+      setup((err) => {
+        if (err) {
+          return done(err)
+        }
+
+        erm.serve(app, db.models.Customer, options)
+
+        server = app.listen(testPort, done)
+      })
+    })
+
+    afterEach((done) => {
+      options.postProcess.reset()
+      dismantle(app, server, done)
+    })
+
+    it('GET /Customer 200', (done) => {
+      request.get({
+        url: `${testUrl}/api/v1/Customer`,
+        json: true
+      }, (err, res, body) => {
+        assert.ok(!err)
+        assert.equal(res.statusCode, 200)
+        sinon.assert.calledOnce(options.postProcess)
+        let args = options.postProcess.args[0]
+        assert.equal(args.length, 3)
+        assert.deepEqual(args[0].erm.result, [])
+        assert.equal(args[0].erm.statusCode, 200)
+        assert.equal(typeof args[2], 'function')
+        done()
+      })
+    })
+  })
+
+  describe('postProcess (async outputFn)', () => {
+    let app = createFn()
+    let server
+    let options = {
+      outputFn: (req, res) => {
+        if (app.isRestify) {
+          res.send(200)
+        } else {
+          res.sendStatus(200)
+        }
+
+        return Promise.resolve()
+      },
       postProcess: sinon.spy((req, res, next) => {
         next()
       }),